+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog
 0.x.x (TBD)
 ===========
 
-* ...
+* Added timeField field for the Elasticsearch target to allow the alert to change its state
 
 Changes
 -------
@@ -25,7 +25,6 @@ Changes
 * Added support for Elasticsearch alert condition
 * Added support for using gridPos for dashboard panels
 * Added support for Humio Data Source. (https://grafana.com/grafana/plugins/humio-datasource/)
-* Added timeField field for the Elasticsearch target to allow the alert to change its state
 
 
 Changes