+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 ===========
 
 * Added timeField field for the Elasticsearch target to allow the alert to change its state
+* Added nameFilter field for the AlertList panel
 
 Changes
 -------